# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ironhack AI Course Copilot API",
    "version": "0.3.0"
  },
  "paths": {
...
    "/api/session": {
      "post": {
        "summary": "New Session",
        "operationId": "new_session_api_session_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response New Session Api Session Post"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/ask": {
      "post": {
        "summary": "Ask",
        "description": "One turn. Creates a session if the client did not supply one.\n\n`rehydrated` is in the response on purpose: it is how the frontend (and we, during\nthe spike) can see when an answer came from a Copilot rebuilt from stored state\nrather than one that was already live. That flag is the experiment.",
        "operationId": "ask_api_ask_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/lessons/{lesson_id}/notes": {
      "get": {
        "summary": "Get Notes",
        "description": "Study notes for one lesson, as markdown.\n\nMarkdown rather than HTML because the frontend should own presentation \u2014 the same\nreason `build_citation` hands over a label and a URL rather than a rendered link.",
        "operationId": "get_notes_api_lessons__lesson_id__notes_get",
        "parameters": [
          {
            "name": "lesson_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Lesson Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Notes Api Lessons  Lesson Id  Notes Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/lessons/{lesson_id}/notes.pdf": {
      "get": {
        "summary": "Get Notes Pdf",
        "description": "The same PDF the Streamlit download button produces, from the same code.\n\n`src/pdf.py` is shared rather than reimplemented \u2014 two renderers would drift and\nstudents would get different documents depending on which frontend they used.",
        "operationId": "get_notes_pdf_api_lessons__lesson_id__notes_pdf_get",
        "parameters": [
          {
            "name": "lesson_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Lesson Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/session/{session_id}/scope": {
      "post": {
        "summary": "Set Scope",
        "description": "Narrow the search to a lesson or a week for the rest of the conversation.\n\nSet on the retrieval side rather than worded into the question \u2014 see SearchScope.\nA scoped refusal says \"not in THIS lesson\", which is a different fact from \"not in\nthe course\", and the frontend should show the filter that caused it.",
        "operationId": "set_scope_api_session__session_id__scope_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScopeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Set Scope Api Session  Session Id  Scope Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/session/{session_id}/quiz": {
      "post": {
        "summary": "Quiz",
        "description": "Generate a scored quiz on a topic, honouring the session's scope.\n\nCalls the tool directly rather than asking the agent to pick it. The agent route\nworks but costs an extra model call to decide something the button already decided.",
        "operationId": "quiz_api_session__session_id__quiz_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuizRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Quiz Api Session  Session Id  Quiz Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/session/{session_id}/reset": {
      "post": {
        "summary": "Reset",
        "operationId": "reset_api_session__session_id__reset_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Reset Api Session  Session Id  Reset Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/session/{session_id}/evict": {
      "post": {
        "summary": "Evict",
        "description": "Drop the live Copilot, keep the conversation.\n\nThe most important endpoint here. Call it between two turns and the next answer has\nto come from a Copilot rebuilt out of stored state \u2014 which is exactly what happens\nwhen a follow-up lands on a replica that never saw the first question. If the\nfollow-up still resolves, horizontal scaling is unblocked.",
        "operationId": "evict_api_session__session_id__evict_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Evict Api Session  Session Id  Evict Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/session/{session_id}": {
      "get": {
        "summary": "Get Session",
        "operationId": "get_session_api_session__session_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Session Api Session  Session Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AskRequest": {
        "properties": {
          "question": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Question"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "language": {
            "type": "string",
            "title": "Language",
            "default": "auto"
          }
        },
        "type": "object",
        "required": [
          "question"
        ],
        "title": "AskRequest",
        "description": "`language` is \"auto\", \"en\" or \"es\".\n\nAuto is the default and is handled by the system prompt, which answers in whatever\nlanguage the student wrote in. An explicit choice appends an instruction to the\nquestion sent to the agent, leaving the question shown in the transcript unchanged \u2014\nsame approach the Streamlit app takes, so both frontends behave identically."
      },
      "AskResponse": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "answer": {
            "type": "string",
            "title": "Answer"
          },
          "citations": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Citations"
          },
          "related_notebooks": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Related Notebooks",
            "default": []
          },
          "tools_used": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tools Used",
            "default": []
          },
          "elapsed_seconds": {
            "type": "number",
            "title": "Elapsed Seconds"
          },
          "rehydrated": {
            "type": "boolean",
            "title": "Rehydrated",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "answer",
          "citations",
          "elapsed_seconds"
        ],
        "title": "AskResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ScopeRequest": {
        "properties": {
          "lesson_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lesson Id"
          },
          "week": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Week"
          }
        },
        "type": "object",
        "title": "ScopeRequest",
        "description": "Empty body clears the scope, which is how the UI turns the filter off."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}